♻️ refactor(websearch): centralize lynx header generation
- add `_get_lynx_useragent` function to generate random lynx headers
- introduce `HEADER_LYNX` constant for standardized header usage
- replace inline lynx header definitions with `HEADER_LYNX` constant across modules
