+++ src/toolregistry/hub/websearch/websearch.py
@@ -1,3 +1,4 @@
+import random
 import re
 import sys
 import unicodedata
@@ -13,7 +14,8 @@ _UNABLE_TO_FETCH_TITLE = "Unable to fetch title"
 
 if sys.version_info < (3, 9):
     HEADERS_DEFAULT = {
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/122.0.0.0 Safari/537.36 Edg/122.0.0.0"
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/122.0.0.0 Safari/537.36 Edg/122.0.0.0",
+        "Accept": "*/*",
     }
 else:
     from fake_useragent import UserAgent
@@ -22,6 +24,34 @@ else:
 TIMEOUT_DEFAULT = 10.0
 
 
+def _get_lynx_useragent():
+    """
+    Generates a random user agent string mimicking the format of various software versions.
+
+    The user agent string is composed of:
+    - Lynx version: Lynx/x.y.z where x is 2-3, y is 8-9, and z is 0-2
+    - libwww version: libwww-FM/x.y where x is 2-3 and y is 13-15
+    - SSL-MM version: SSL-MM/x.y where x is 1-2 and y is 3-5
+    - OpenSSL version: OpenSSL/x.y.z where x is 1-3, y is 0-4, and z is 0-9
+
+    Returns:
+        str: A randomly generated user agent string.
+    """
+    lynx_version = (
+        f"Lynx/{random.randint(2, 3)}.{random.randint(8, 9)}.{random.randint(0, 2)}"
+    )
+    libwww_version = f"libwww-FM/{random.randint(2, 3)}.{random.randint(13, 15)}"
+    ssl_mm_version = f"SSL-MM/{random.randint(1, 2)}.{random.randint(3, 5)}"
+    openssl_version = (
+        f"OpenSSL/{random.randint(1, 3)}.{random.randint(0, 4)}.{random.randint(0, 9)}"
+    )
+    return f"{lynx_version} {libwww_version} {ssl_mm_version} {openssl_version}"
+
+
+HEADER_LYNX = {
+    "User-Agent": _get_lynx_useragent(),
+    "Accept": "*/*",
+}
 
 class _WebSearchEntryGeneral(dict):
     def __init__(self, **data):
